CleanUp - don't work replay blocks
