+++ rspace/src/main/scala/coop/rchain/rspace/history/HistoryRepository.scala
@@ -5,7 +5,7 @@ import cats.effect.Concurrent
 import cats.syntax.all._
 import coop.rchain.metrics.Span
 import coop.rchain.rspace.hashing.Blake2b256Hash
-import coop.rchain.rspace.history.instances.RadixHistory7
+import coop.rchain.rspace.history.instances.RadixHistory
 import coop.rchain.rspace.serializers.ScodecSerialize.{DatumB, JoinsB, WaitingContinuationB}
 import coop.rchain.rspace.state.instances.{RSpaceExporterStore, RSpaceImporterStore}
 import coop.rchain.rspace.state.{RSpaceExporter, RSpaceImporter}
@@ -73,12 +73,8 @@ object HistoryRepositoryInstances {
     for {
       currentRoot <- rootsRepository.currentRoot()
       // History store
-//      historyStore = HistoryStoreInstances.historyStore[F](historyKeyValueStore)
-//      history      = HistoryInstances.merging(currentRoot, historyStore)
-//      history = RadixHistory3(currentRoot, TrieMap[ByteVector, ByteVector])
-
-      historyStore = new RadixStore(historyKeyValueStore)
-      history      = RadixHistory7(currentRoot, historyStore)
+      historyStore = new RadixStore[F](historyKeyValueStore)
+      history      <- RadixHistory(currentRoot, historyStore)
 
       // Cold store
       coldStore = ColdStoreInstances.coldStore[F](coldKeyValueStore)
