+++ rspace/src/main/scala/coop/rchain/rspace/history/RadixTree.scala
@@ -1,319 +1,581 @@
 package coop.rchain.rspace.history
 
+import cats.Parallel
+import cats.effect.Sync
 import cats.syntax.all._
-import coop.rchain.rspace.hashing.Blake2b256Hash
-import scodec.Codec
 import scodec.bits.ByteVector
 
 import scala.annotation.tailrec
 import scala.collection.concurrent.TrieMap
+import scala.language.higherKinds
 
+/**
+  * Radix Tree implementation
+  */
 object RadixTree {
 
   /**
-    * Radix Tree structure (Node and Leaf)
+    * Child Node structure (Empty,Leaf and NodePointer)
+    * There is only one type of element in Tree [[Node]] = Vector[Child]
     */
-  sealed trait TreeR
-  final case class NodeR(prefix: ByteVector, refs: Vector[ByteVector]) extends TreeR
-  final case class LeafR(prefix: ByteVector, value: ByteVector)        extends TreeR
+  sealed trait Child
+  final case object EmptyChild                                                     extends Child
+  final case class Leaf(prefix: ByteVector, value: ByteVector, varLength: Boolean) extends Child
+  final case class NodePtr(prefix: ByteVector, ptr: ByteVector)                    extends Child
 
+  type Node = Vector[Child]
+
+  /**
+    * number of children always constant
+    */
+  val numChildren = 256
+
+  /**
+    * Empty node consists only of [[EmptyChild]]'s
+    */
+  val emptyNode: Node = (0 until numChildren).map(_ => EmptyChild).toVector
+
+  /**
+    * Binary codecs for serializing/deserializing Node in Radix tree
+    * Child coding structure:
+    *   EmptyChild                   - Empty (not encode)
+
+    *   Leaf(prefix,value,varLength)
+    *           if(varLength) -> [child number] [second byte] [prefix0]..[prefixM] [sizeValue=N] [value0]..[valueN]
+    *           else          -> [child number] [second byte] [prefix0]..[prefixM] [value0]..[value31]
+    *                            [second byte] -> bit7 = 0 (Leaf identifier)
+    *                                             bit6 = varLength flag (0 - value have length 32 bytes
+    *                                                                    1  - value have variable length 0..255 bytes)
+    *                                             bit5 = reserv
+    *                                             bit4..bit0 - prefix length = M (from 0 to 31)
+    *
+    *   NodePtr(prefix,ptr)   -> [child number] [second byte] [prefix0]..[prefixM] [ptr0]..[ptr31]
+    *                            [second byte] -> bit7 = 1 (NodePtr identifier)
+    *                                             bit6 = reserv
+    *                                             bit5 = reserv
+    *                                             bit4..bit0 - prefix length = M (from 0 to 31)
+    */
   object codecs {
-    import scodec.codecs._
+    def encode(node: Node): Array[Byte] = {
+      //calculating size of buffer
+      val sizeNode = node.size
 
-    val codecByteVector: Codec[ByteVector] = variableSizeBytes(uint8, bytes)
+      @tailrec
+      def loopSize(index: Int, size: Int): Int =
+        if (sizeNode > index)
+          node(index) match {
+            case EmptyChild => loopSize(index + 1, size)
 
-    // Binary codecs for Tree
+            case Leaf(leafPrefix, value, varLength) =>
+              val sizePrefix = leafPrefix.size.toInt
+              val sizeValue  = value.size.toInt
+              val sizeLength = if (varLength) {
+                assert(
+                  value.size <= 255,
+                  "error during deserialization (size of value more than 255)"
+                )
+                1
+              } else {
+                assert(
+                  sizeValue == 32,
+                  "error during deserialization (size of leafValue not equal 32)"
+                )
+                0
+              }
+              loopSize(index + 1, size + 2 + sizePrefix + +sizeLength + sizeValue)
 
-    val codecNodeR: Codec[NodeR] =
-      (codecByteVector :: vectorOfN(provide(256), codecByteVector)).as[NodeR]
+            case NodePtr(ptrPrefix, ptr) =>
+              val sizePrefix = ptrPrefix.size.toInt
+              assert(sizePrefix <= 31, "error during deserialization (size of prefix more than 31)")
+              val sizePtr = ptr.length.toInt
+              assert(sizePtr == 32, "error during deserialization (size of ptrPrefix not equal 32)")
+              loopSize(index + 1, size + 2 + sizePrefix + sizePtr)
+          } else size
 
-    val codecLeafR: Codec[LeafR] = (codecByteVector :: codecByteVector).as[LeafR]
+      val size = loopSize(0, 0)
+      //put in the buffer
 
-    val codecTreeR: DiscriminatorCodec[TreeR, Int] =
-      discriminated[TreeR]
-        .by(uint2)
-        .subcaseP(tag = 0) {
-          case s: LeafR => s
-        }(codecLeafR)
-        .subcaseP(tag = 1) {
-          case n: NodeR => n
-        }(codecNodeR)
+      val arr = new Array[Byte](size)
+      @tailrec
+      def loopFill(numChild: Int, idx0: Int): Array[Byte] =
+        if (idx0 == size) arr
+        else {
+          node(numChild) match {
+            case EmptyChild => loopFill(numChild + 1, idx0)
+
+            case Leaf(prefix, value, varLength) =>
+              arr(idx0) = numChild.toByte
+              val idxSecondByte   = idx0 + 1
+              val idxPrefixStart  = idxSecondByte + 1
+              val prefixSize: Int = prefix.size.toInt
+              if (varLength) {
+                //01000000b | (size & 00011111b)
+                arr(idxSecondByte) = (0x40 | (prefixSize & 0x1F)).toByte //second byte - type and size of prefix
+                for (i <- 0 until prefixSize) arr(idxPrefixStart + i) = prefix(i.toLong)
+                val idxValueSize = idxPrefixStart + prefixSize
+                val valueSize    = value.size.toInt
+                arr(idxValueSize) = valueSize.toByte
+                val idxValueStart = idxValueSize + 1
+                for (i <- 0 until valueSize) arr(idxValueStart + i) = value(i.toLong)
+                loopFill(numChild + 1, idxValueStart + valueSize)
+              } else {
+                //size & 00011111b
+                arr(idxSecondByte) = (prefixSize & 0x1F).toByte //second byte - type and size of prefix
+                for (i <- 0 until prefixSize) arr(idxPrefixStart + i) = prefix(i.toLong)
+                val idxValueStart = idxPrefixStart + prefixSize
+                for (i <- 0 until 32) arr(idxValueStart + i) = value(i.toLong)
+                loopFill(numChild + 1, idxValueStart + 32)
+              }
+            case NodePtr(prefix, ptr) =>
+              arr(idx0) = numChild.toByte
+              val idxSecondByte   = idx0 + 1
+              val prefixSize: Int = prefix.size.toInt
+              //10000000b | (size & 00011111b)
+              arr(idxSecondByte) = (0x80 | (prefixSize & 0x1F)).toByte //second byte - type and size of prefix
+              val idxPrefixStart = idxSecondByte + 1
+              for (i <- 0 until prefixSize) arr(idxPrefixStart + i) = prefix(i.toLong)
+              val idxPtrStart = idxPrefixStart + prefixSize
+              for (i <- 0 until 32) arr(idxPtrStart + i) = ptr(i.toLong)
+              loopFill(numChild + 1, idxPtrStart + 32)
+          }
+        }
+      loopFill(0, 0)
+    }
+
+    def decode(arr: Array[Byte]): Node = {
+      val maxSize = arr.length
+      @tailrec //todo it's need?
+      def loop(idx0: Int, node: Node): Node =
+        if (idx0 == maxSize) node
+        else {
+          val (idx0Next, nodeNext) = try {
+            val childNum: Int = byteToInt(arr(idx0))
+            assert(
+              node(childNum) == EmptyChild,
+              "error during deserialization (wrong child number)"
+            )
+            val idx1       = idx0 + 1
+            val secondByte = arr(idx1)
+            assert((secondByte & 0x20) == 0x00, "error during deserialization (wrong type format)")
+
+            val prefixSize: Int = secondByte & 0x1F
+            val prefix          = new Array[Byte](prefixSize)
+            val idxPrefixStart  = idx1 + 1
+            for (i <- 0 until prefixSize) prefix(i) = arr(idxPrefixStart + i)
+
+            if ((secondByte & 0x80) == 0x00) {   //Leaf
+              if ((secondByte & 0x40) != 0x00) { //leafValue have variable length
+                val idxleafValueSize   = idxPrefixStart + prefixSize
+                val leafValueSize: Int = byteToInt(arr(idxleafValueSize))
+                val leafValue          = new Array[Byte](leafValueSize)
+                val idxleafValueStart  = idxleafValueSize + 1
+                for (i <- 0 until leafValueSize) leafValue(i) = arr(idxleafValueStart + i)
+                val idx0Next = idxleafValueStart + leafValueSize
+                val nodeNext =
+                  node.updated(
+                    childNum,
+                    Leaf(ByteVector(prefix), ByteVector(leafValue), varLength = true)
+                  )
+                (idx0Next, nodeNext)
+              } else { //leafValue is 32 byte ptr
+                val leafValue         = new Array[Byte](32)
+                val idxleafValueStart = idxPrefixStart + prefixSize
+                for (i <- 0 until 32) leafValue(i) = arr(idxleafValueStart + i)
+                val idx0Next = idxleafValueStart + 32
+                val nodeNext =
+                  node.updated(
+                    childNum,
+                    Leaf(ByteVector(prefix), ByteVector(leafValue), varLength = false)
+                  )
+                (idx0Next, nodeNext)
+              }
+            } else { //NodePtr
+              val ptr         = new Array[Byte](32)
+              val idxPtrStart = idxPrefixStart + prefixSize
+              for (i <- 0 until 32) ptr(i) = arr(idxPtrStart + i)
+              val idx0Next = idxPtrStart + 32
+              val nodeNext =
+                node.updated(
+                  childNum,
+                  NodePtr(ByteVector(prefix), ByteVector(ptr))
+                )
+              (idx0Next, nodeNext)
+            }
+          } catch {
+            case _: Exception =>
+              assert(assertion = false, "error during deserialization (out of range)")
+              (maxSize, emptyNode)
+          }
+          loop(idx0Next, nodeNext)
+        }
+      loop(0, emptyNode)
+    }
   }
 
-  def commonPrefix(b1: ByteVector, b2: ByteVector) = {
+  /**
+    * Find the common part of b1 and b2.
+    * Return (Common part , rest of b1, rest of b2).
+    */
+  def commonPrefix(b1: ByteVector, b2: ByteVector): (ByteVector, ByteVector, ByteVector) = {
     @tailrec
-    def loop(
+    def go(
         common: ByteVector,
         l: ByteVector,
         r: ByteVector
     ): (ByteVector, ByteVector, ByteVector) =
-//      assert(l.size >= r.size, s"Key ${b1.toHex} shorter then stored prefix ${b2.toHex}.")
       if (r.isEmpty) {
         (common, l, r)
       } else {
         val lHead = l.head
         val rHead = r.head
-        if (lHead == rHead) loop(common :+ lHead, l.tail, r.tail)
+        if (lHead == rHead) go(common :+ lHead, l.tail, r.tail)
         else (common, l, r)
       }
-
-    loop(ByteVector.empty, b1, b2)
+    go(ByteVector.empty, b1, b2)
   }
 
-  def hashFn(bs: ByteVector): Blake2b256Hash =
-    Blake2b256Hash.create(bs)
-
-  def hashTree(tree: TreeR) = {
-    val newRefsSer = codecs.codecTreeR.encode(tree).require.toByteVector
-    (hashFn(newRefsSer), newRefsSer)
+  /**
+    * Hashing serial data.
+    * Return Blake2b256 hash of input data
+    */
+  def hashNode(node: Node): (ByteVector, Array[Byte]) = {
+    import coop.rchain.crypto.hash.Blake2b256
+    val bytes = codecs.encode(node)
+    (ByteVector(Blake2b256.hash(bytes)), bytes)
   }
 
-  // Default values for empty Tree/Node
+  def byteToInt(b: Byte): Int = b & 0xff
 
-  val emptyTreeRefs: Vector[ByteVector] = (0 to 255).map(_ => ByteVector.empty).toVector
+  class RadixTreeImpl[F[_]: Sync: Parallel](store: RadixStore[F]) {
 
-  val emptyTree: TreeR = {
-    val emptyVec = ByteVector.empty
-    NodeR(prefix = emptyVec, refs = emptyTreeRefs)
-  }
+    /**
+      * Load and decode serializing data from KVDB.
+      */
+    private def loadNodeFromStore(nodePtr: ByteVector): F[Option[Node]] =
+      store.get(Seq(nodePtr)).map(_.head.map(v => codecs.decode(v)))
 
-  class RadixTreeImpl(store: TrieMap[ByteVector, ByteVector]) {
+    /**
+      * Cache for storing read and decoded nodes.
+      * In cache load kv-pair (hash, node).
+      * Where hash  - Blake2b256Hash of serializing nodes data,
+      *       node - deserialized data of this node.
+      */
+    private val cacheR: TrieMap[ByteVector, Node] = TrieMap.empty
 
-    import codecs._
+    /**
+      * Load one node from [[cacheR]].
+      * If there is no such record in cache - load and decode from KVDB, then save to cacheR.
+      * If there is no such record in KVDB - execute assert (if set noAssert flag - return emptyNode).
+      */
+    def loadNode(nodePtr: ByteVector, noAssert: Boolean = false): F[Node] =
+      for {
+        maybeCacheValue <- Sync[F].delay(cacheR.get(nodePtr))
+        cacheValue = maybeCacheValue match {
+          case None =>
+            loadNodeFromStore(nodePtr).map {
+              case None =>
+                assert(noAssert, s"Missing node in database. ptr=${nodePtr.toHex}")
+                emptyNode
+              case Some(storeNode) =>
+                cacheR.update(nodePtr, storeNode)
+                storeNode
+            }
+          case Some(v) => Sync[F].pure(v)
+        }
+        result <- cacheValue
+      } yield result
 
-    def toInt(b: Byte): Int = b & 0xff
+    /**
+      * Clear [[cacheR]] (cache for storing read nodes).
+      */
+    def clearReadCache(): Unit = cacheR.clear()
 
-    def loadRefOpt(treeKey: ByteVector): Option[TreeR] = {
-      val nodeData = store.get(treeKey)
-      nodeData.map { data =>
-        codecTreeR.decode(data.toBitVector).require.value
+    /**
+      * Cache for storing serializing nodes. For subsequent unloading in KVDB
+      * In cache store kv-pair (hash, bytes).
+      * Where hash -  Blake2b256Hash of bytes,
+      *       bytes - serializing data of nodes.
+      */
+    private val cacheW: TrieMap[ByteVector, Array[Byte]] = TrieMap.empty
+
+    /**
+      * Serializing and hashing one [[Node]].
+      * Serializing data load in [[cacheW]].
+      * If detected collision with older cache data - executing assert
+      */
+    def saveNode(node: Node): ByteVector = {
+      val (hash, bytes) = hashNode(node)
+      cacheR.get(hash) match { // collision alarm
+        case Some(v) =>
+          assert(
+            v == node,
+            s"collision in cache (record with key = ${hash.toHex} has already existed)"
+          )
+        case None => cacheR.update(hash, node)
       }
-    }
-
-    def loadRef(treeKey: ByteVector): TreeR =
-      loadRefOpt(treeKey).get
-
-    def saveRef(tree: TreeR): Blake2b256Hash = {
-      val (hash, bytes) = hashTree(tree)
-      store.update(hash.bytes, bytes)
+      cacheW.update(hash, bytes)
       hash
     }
 
-    def loadRefWithDefault(hash: Blake2b256Hash): TreeR =
-      loadRefOpt(hash.bytes).getOrElse(emptyTree)
+    /**
+      * Save all cacheW to KVDB
+      * If detected collision with older KVDB data - execute assert
+      */
+    def commit(): F[Unit] = {
+      val kvPairs = cacheW.toList
+      for {
+        ifAbsent          <- store.contains(kvPairs.map(_._1))
+        kvIfAbsent        = kvPairs zip ifAbsent
+        kvExist           = kvIfAbsent.filter(_._2).map(_._1)
+        valueExistInStore <- store.get(kvExist.map(_._1))
+        kvvExist          = kvExist zip valueExistInStore.map(_.getOrElse(Array.empty))
+        kvCollision       = kvvExist.filter(kvv => !(kvv._1._2 sameElements kvv._2)).map(_._1)
+        kvAbsent = if (kvCollision.isEmpty) kvIfAbsent.filterNot(_._2).map(_._1)
+        else {
+          assert(
+            assertion = false,
+            s"${kvCollision.length} collisions in KVDB (first collision with key = ${kvCollision.head._1.toHex})"
+          )
+          List.empty
+        }
+        _ <- store.put(kvAbsent)
+      } yield ()
+    }
 
     /**
-      * Read key with prefix
+      * Clear [[cacheW]] (cache for storing data to write in KVDB).
       */
-    @tailrec
-    final def read(tree: TreeR, prefix: ByteVector): Option[ByteVector] =
-      tree match {
-        case NodeR(nPrefix, nRefs) =>
-          assert(prefix.size > nPrefix.size, "Key is shorter then stored prefix.")
+    def clearWriteCache(): Unit = cacheW.clear()
 
-          val (_, prefixRest, nPrefixRest) = commonPrefix(prefix, nPrefix)
+    /**
+      * Read leaf data with prefix. If data not found, returned [[None]]
+      */
+    final def read(startNode: Node, startPrefix: ByteVector): F[Option[ByteVector]] = {
+      type Params = (Node, ByteVector)
+      def go(params: Params): F[Either[Params, Option[ByteVector]]] =
+        params match {
+          case (_, ByteVector.empty) => Sync[F].pure(None.asRight) //Not found
+          case (curNode, prefix) =>
+            curNode(byteToInt(prefix.head)) match {
+              case EmptyChild => Sync[F].pure(None.asRight) //Not found
 
-          if (nPrefixRest.isEmpty) {
-            // Found prefix
-            val idx        = toInt(prefixRest.head)
-            val prefixTail = prefixRest.tail
-            val ref        = nRefs(idx)
-            if (ref.isEmpty) {
-              none
-            } else {
-              val node = loadRef(ref)
-              read(node, prefixTail)
+              case Leaf(leafPrefix, value, _) =>
+                if (leafPrefix == prefix.tail) Sync[F].pure(value.some.asRight) //Happy end
+                else Sync[F].pure(None.asRight)                                 //Not found
+
+              case NodePtr(ptrPrefix, ptr) =>
+                val (_, prefixRest, ptrPrefixRest) = commonPrefix(prefix.tail, ptrPrefix)
+                if (ptrPrefixRest.isEmpty) loadNode(ptr).map(n => (n, prefixRest).asLeft) //Deeper
+                else Sync[F].pure(None.asRight)                                           //Not found
             }
-          } else {
-            // Key not found
-            none
-          }
+        }
+      Sync[F].tailRecM(startNode, startPrefix)(go)
+    }
 
-        case LeafR(nPrefix, value) =>
-          if (nPrefix == prefix) value.some
-          else none
+    /**
+      * Create node from [[Child]].
+      * If child is NodePtr and prefix is empty - load node
+      */
+    private def createNodeFromChild(child: Child): F[Node] =
+      child match {
+        case EmptyChild => emptyNode.pure
+        case Leaf(leafPrefix, leafValue, varLength) =>
+          assert(
+            leafPrefix.nonEmpty,
+            "Impossible to create a node. LeafPrefix should be non empty."
+          )
+          emptyNode
+            .updated(byteToInt(leafPrefix.head), Leaf(leafPrefix.tail, leafValue, varLength))
+            .pure
+        case NodePtr(nodePtrPrefix, ptr) =>
+          if (nodePtrPrefix.isEmpty) loadNode(ptr)
+          else
+            emptyNode
+              .updated(byteToInt(nodePtrPrefix.head), NodePtr(nodePtrPrefix.tail, ptr))
+              .pure
       }
 
     /**
-      * Write value on a key with prefix
+      * Optimize and save Node, create Child from this Node
       */
-    def write(tree: TreeR, prefix: ByteVector, value: Option[ByteVector]): TreeR =
-      value.map(update(tree, prefix, _)).getOrElse(delete(tree, prefix))
+    private def saveNodeAndCreateChild(
+        node: Node,
+        childPrefix: ByteVector,
+        optimization: Boolean = true
+    ): Child =
+      if (optimization) {
+        val nonEmptyChildren = node.iterator.filter(_ != EmptyChild).take(2).toList
+        nonEmptyChildren.size match {
+          case 0 => EmptyChild //all children are empty
+          case 1 => //only one child is not empty - merge child and parent nodes
+            val childIdx = node.indexOf(nonEmptyChildren.head)
+            nonEmptyChildren.head match {
+              case EmptyChild => EmptyChild
+              case Leaf(leafPrefix, value, varLength) =>
+                Leaf(childPrefix ++ ByteVector(childIdx) ++ leafPrefix, value, varLength)
+              case NodePtr(nodePtrPrefix, ptr) =>
+                NodePtr(childPrefix ++ ByteVector(childIdx) ++ nodePtrPrefix, ptr)
+            }
+          case 2 => //2 or more children are not empty
+            NodePtr(childPrefix, saveNode(node))
+        }
+      } else NodePtr(childPrefix, saveNode(node))
 
     /**
-      * Insert or update key with prefix
-      *
-      * NOTE: Not tail recursive but it should be safe because tree can be only _prefix_ length deep.
+      * Save new leaf value to this part of tree (start from curChild).
+      * Rehash and save all depend node to cacheW. Return updated curChild.
+      * If exist leaf with same prefix but different value - update leaf value.
+      * If exist leaf with same prefix and same value - return [[None]].
       */
-    def update(tree: TreeR, prefix: ByteVector, value: ByteVector): TreeR =
-      tree match {
-        case NodeR(nPrefix, nRefs) =>
-          assert(prefix.size > nPrefix.size, "Storing shorter prefix then already stored.")
+    def update(
+        curChild: Child,
+        insPrefix: ByteVector,
+        insValue: ByteVector,
+        varLength: Boolean
+    ): F[Option[Child]] =
+      curChild match {
+        case EmptyChild =>
+          (Leaf(insPrefix, insValue, varLength): Child).some.pure //update EmptyChild to Leaf
 
-          val (commPrefix, prefixNew, nPrefixNew) = commonPrefix(prefix, nPrefix)
-
-          if (nPrefixNew.isEmpty) {
-            // Insert value in existing refs
-            val (prefixNewIdx, prefixNewTail) = (toInt(prefixNew.head), prefixNew.tail)
-            val ref                           = nRefs(prefixNewIdx)
-            val node =
-              if (ref == ByteVector.empty) {
-                // No existing node, insert leaf
-                LeafR(prefix.tail, value)
-              } else {
-                // Found node by prefix in refs, load update recursively
-                val idxTree = loadRef(ref)
-                val newTree = update(idxTree, prefixNewTail, value)
-                newTree
-              }
-            val newRef  = saveRef(node)
-            val newRefs = nRefs.updated(prefixNewIdx, newRef.bytes)
-            NodeR(commPrefix ++ nPrefixNew, newRefs)
-          } else {
-            // Move existing refs, create sub-node for existing refs and leaf for insert
-            val (nPrefixNewIdx, nPrefixNewTail) = (toInt(nPrefixNew.head), nPrefixNew.tail)
-            val thisNode                        = NodeR(nPrefixNewTail, nRefs)
-            val thisNodeRef                     = saveRef(thisNode)
-
-            val (prefixNewIdx, prefixNewTail) = (toInt(prefixNew.head), prefixNew.tail)
-            val newNode                       = LeafR(prefixNewTail, value)
-            val newNodeRef                    = saveRef(newNode)
-
-            val newRefs = emptyTreeRefs
-              .updated(nPrefixNewIdx, thisNodeRef.bytes)
-              .updated(prefixNewIdx, newNodeRef.bytes)
-
-            NodeR(commPrefix, newRefs)
+        case Leaf(leafPrefix, leafValue, flag) =>
+          assert(leafPrefix.size == insPrefix.size, "All Radix keys should be same length")
+          if (leafPrefix == insPrefix) {
+            if (insValue == leafValue) none[Child].pure
+            else (Leaf(insPrefix, insValue, varLength): Child).some.pure
+          } //update Leaf
+          else {
+            // Create child node, insert existing and new leaf in this node
+            // intentionally not recursive for speed up
+            val (commPrefix, insPrefixRest, leafPrefixRest) = commonPrefix(insPrefix, leafPrefix)
+            val newNode = emptyNode
+              .updated(byteToInt(leafPrefixRest.head), Leaf(leafPrefixRest.tail, leafValue, flag))
+              .updated(byteToInt(insPrefixRest.head), Leaf(insPrefixRest.tail, insValue, varLength))
+            saveNodeAndCreateChild(newNode, commPrefix, optimization = false).some.pure
           }
 
-        case LeafR(nPrefix, nValue) =>
-          if (nPrefix == prefix) {
-            // Update value
-            LeafR(prefix, value)
+        case NodePtr(ptrPrefix, ptr) =>
+          assert(ptrPrefix.size < insPrefix.size, "Radix key should be longer than NodePtr key")
+          val (commPrefix, insPrefixRest, ptrPrefixRest) = commonPrefix(insPrefix, ptrPrefix)
+          if (ptrPrefixRest.isEmpty) {
+            val (childChildIdx, childInsPrefix) =
+              (byteToInt(insPrefixRest.head), insPrefixRest.tail)
+            //add new node to existing child node
+            for {
+              childNode     <- loadNode(ptr)
+              childChildOpt <- update(childNode(childChildIdx), childInsPrefix, insValue, varLength) //deeper
+              returnedChild = childChildOpt match {
+                case None => none
+                case Some(childChild) =>
+                  val updatedChildNode = childNode.updated(childChildIdx, childChild)
+                  saveNodeAndCreateChild(updatedChildNode, commPrefix, optimization = false).some
+              }
+            } yield returnedChild
           } else {
-            val (commPrefix, prefixNew, nPrefixNew) = commonPrefix(prefix, nPrefix)
-
-            // Create sub-node, insert existing and new leaf
-            val (nPrefixNewIdx, nPrefixNewTail) = (toInt(nPrefixNew.head), nPrefixNew.tail)
-            val thisNode                        = LeafR(nPrefixNewTail, nValue)
-            val thisNodeRef                     = saveRef(thisNode)
-
-            val (prefixNewIdx, prefixNewTail) = (toInt(prefixNew.head), prefixNew.tail)
-            val newNode                       = LeafR(prefixNewTail, value)
-            val newNodeRef                    = saveRef(newNode)
-
-            val newRefs = emptyTreeRefs
-              .updated(nPrefixNewIdx, thisNodeRef.bytes)
-              .updated(prefixNewIdx, newNodeRef.bytes)
-
-            NodeR(commPrefix, newRefs)
+            // Create child node, insert existing Ptr and new leaf in this node
+            val newNode = emptyNode
+              .updated(byteToInt(ptrPrefixRest.head), NodePtr(ptrPrefixRest.tail, ptr))
+              .updated(byteToInt(insPrefixRest.head), Leaf(insPrefixRest.tail, insValue, varLength))
+            saveNodeAndCreateChild(newNode, commPrefix, optimization = false).some.pure
           }
       }
 
     /**
-      * Delete key with prefix
-      *
-      * NOTE: Not tail recursive but it should be safe because tree can be only _prefix_ length deep.
+      * Delete leaf value from this part of tree (start from curChild).
+      * Rehash and save all depend node to cacheW. Return updated curChild.
+      * If not found leaf with  delPrefix - return [[None]].
       */
-    def delete(tree: TreeR, prefix: ByteVector): TreeR =
-      tree match {
-        case existingNode @ NodeR(nPrefix, nRefs) =>
-          assert(prefix.size > nPrefix.size, "Storing shorter prefix then already stored.")
+    def delete(curChild: Child, delPrefix: ByteVector): F[Option[Child]] =
+      curChild match {
+        case EmptyChild => none[Child].pure //Not found
 
-          val (commPrefix, prefixNew, nPrefixNew) = commonPrefix(prefix, nPrefix)
+        case Leaf(leafPrefix, _, _) =>
+          if (leafPrefix == delPrefix) (EmptyChild: Child).some.pure //Happy end
+          else none[Child].pure                                      //Not found
 
-          if (nPrefixNew.isEmpty) {
-            // Match prefix with existing node, search in refs
-            val (prefixNewIdx, prefixNewTail) = (toInt(prefixNew.head), prefixNew.tail)
-            val ref                           = nRefs(prefixNewIdx)
-            val node =
-              if (ref == ByteVector.empty) {
-                // Not found in existing refs
-                emptyTree
-              } else {
-                // Found node by prefix in refs, load and delete recursively
-                val idxTree = loadRef(ref)
-                val newTree = delete(idxTree, prefixNewTail)
-                newTree
+        case NodePtr(ptrPrefix, ptr) =>
+          val (commPrefix, delPrefixRest, ptrPrefixRest) = commonPrefix(delPrefix, ptrPrefix)
+          if (ptrPrefixRest.nonEmpty || delPrefixRest.isEmpty) none[Child].pure //Not found
+          else {
+            val (childChildIdx, childDelPrefix) =
+              (byteToInt(delPrefixRest.head), delPrefixRest.tail)
+            for {
+              childNode     <- loadNode(ptr)
+              childChildOpt <- delete(childNode(childChildIdx), childDelPrefix) //deeper
+              returnedChild = childChildOpt match {
+                case None => none
+                case Some(childChild) =>
+                  saveNodeAndCreateChild(childNode.updated(childChildIdx, childChild), commPrefix).some
               }
-            // Check if node is empty
-            if (node == emptyTree) {
-              // Saving empty node
-              val newRefs = nRefs.updated(prefixNewIdx, ByteVector.empty)
-              if (newRefs == emptyTreeRefs) {
-                // If all refs are empty, return whole empty node
-                emptyTree
-              } else {
-                // New refs are not empty, update in existing node
-                val nonEmptyRefs = newRefs.iterator.filter(_.nonEmpty).take(2).toList
-                if (nonEmptyRefs.size == 1) {
-                  // Only one refs, contract node
-                  val ref           = nonEmptyRefs.head
-                  val idx           = newRefs.indexOf(ref)
-                  val refNode       = loadRef(ref)
-                  val prefixWithIdx = nPrefix ++ ByteVector(idx)
-                  // Update prefix
-                  refNode match {
-                    case NodeR(refPrefix, refRefs) =>
-                      val newPrefix = prefixWithIdx ++ refPrefix
-                      NodeR(newPrefix, refRefs)
-                    case LeafR(refPrefix, refValue) =>
-                      val newPrefix = prefixWithIdx ++ refPrefix
-                      LeafR(newPrefix, refValue)
-                  }
-                } else {
-                  // More then one refs, update existing node
-                  NodeR(nPrefix, newRefs)
+            } yield returnedChild
+          }
+      }
+
+    /**
+      * Parallel processing of HistoryActions in this part of tree (start from curNode).
+      * New data load to [[cacheW]].
+      * Return updated curNode. if no action was taken - return [[None]].
+      */
+    def makeActions(curNode: Node, curActions: List[HistoryAction]): F[Option[Node]] = {
+      val groups = curActions.groupBy(_.key.head).toList
+      val newGroupChildrenF = groups.map {
+        case (groupIdx, groupActions) =>
+          val index = byteToInt(groupIdx)
+          val child = curNode(index)
+          if (groupActions.length == 1) {
+            val newChild = groupActions.head match {
+              case InsertAction(key, hash) =>
+                update(child, ByteVector(key).tail, hash.bytes, varLength = false)
+              case DeleteAction(key) => delete(child, ByteVector(key).tail)
+            }
+            newChild.map((index, _))
+          } else {
+            val notExistInsertAction = groupActions.collectFirst { case _: InsertAction => true }.isEmpty
+            val clearGroupActions =
+              if (child == EmptyChild && notExistInsertAction) groupActions.collect {
+                case v: InsertAction => v
+              } else groupActions
+            if (clearGroupActions.isEmpty) (index, none[Child]).pure
+            else {
+              val newActions = clearGroupActions.map {
+                case InsertAction(key, hash) => InsertAction(key.tail, hash)
+                case DeleteAction(key)       => DeleteAction(key.tail)
+              }
+              for {
+                createdNode <- createNodeFromChild(curNode(index))
+                newNodeOpt  <- makeActions(createdNode, newActions)
+                newChild = newNodeOpt match {
+                  case None          => none
+                  case Some(newNode) => saveNodeAndCreateChild(newNode, ByteVector.empty).some
                 }
-              }
-            } else {
-              // Not empty node, save changes
-              val newRef  = saveRef(node)
-              val newRefs = nRefs.updated(prefixNewIdx, newRef.bytes)
-              NodeR(commPrefix ++ nPrefixNew, newRefs)
+              } yield (index, newChild)
             }
-          } else {
-            // Missing node, nothing to delete
-            existingNode
-          }
-
-        case existingLeaf @ LeafR(nPrefix, _) =>
-          if (nPrefix == prefix) {
-            // Delete value
-            emptyTree
-          } else {
-            existingLeaf
           }
       }
+      for {
+        newGroupChildren <- newGroupChildrenF.parSequence
+        newCurNode = newGroupChildren.foldLeft(curNode) {
+          case (tempNode, (index, newChildOpt)) =>
+            newChildOpt.map(tempNode.updated(index, _)).getOrElse(tempNode)
+        }
+      } yield if (newCurNode != curNode) newCurNode.some else none
+    }
 
     /**
       * Pretty print radix tree
       */
-    final def print(
-        tree: TreeR,
-        indentLevel: Int = 0,
-        prefix: ByteVector = ByteVector.empty,
-        treeKey: ByteVector = ByteVector.empty
-    ): Unit = {
-      val indent         = Seq.fill(indentLevel * 2)(" ").mkString
-      def prn(s: String) = println(s"$indent$s")
-      tree match {
-        case NodeR(nPrefix, nRefs) =>
-          prn(s"${prefix.toHex}${nPrefix.toHex} [${treeKey.toHex}]")
+    final def print(curNode: Node, indentLevel: Int = 0): Unit = {
+      val indent               = Seq.fill(indentLevel * 2)(" ").mkString
+      def prn(s: String): Unit = println(s"$indent$s")
+      curNode.zipWithIndex foreach {
+        case (EmptyChild, _) => Unit
 
-          nRefs.zipWithIndex.foreach {
-            case (ref, idx) =>
-              if (ref.nonEmpty) {
-                val refNode = loadRef(ref)
-                print(refNode, indentLevel + 1, ByteVector(idx), ref)
-              }
-          }
+        case (Leaf(leafPrefix, value, _), idx) =>
+          prn(s"${idx.toHexString.toUpperCase}${leafPrefix.toHex}: ${value.toHex}")
 
-        case LeafR(nPrefix, value) =>
-          prn(s"${prefix.toHex}${nPrefix.toHex}: ${value.toHex} [${treeKey.toHex}]")
+        case (NodePtr(ptrPrefix, ptr), idx) =>
+          prn(s"${idx.toHexString.toUpperCase}${ptrPrefix.toHex} [${ptr.toHex}]")
+          loadNode(ptr).map(print(_, indentLevel + 1))
       }
     }
-
   }
-
 }
